requirements.txt
absl-py==2.1.0
attrs==23.2.0
blinker==1.8.2
certifi==2024.6.2
cffi==1.16.0
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
filelock==3.15.4
Flask==3.0.3
fsspec==2024.6.1
grpcio==1.65.1
h11==0.14.0
idna==3.7
intel-openmp==2021.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
Markdown==3.6
MarkupSafe==2.1.5
mkl==2021.4.0
mpmath==1.3.0
networkx==3.3
numpy==1.26.3
outcome==1.3.0.post0
pillow==10.3.0
pip==23.2.1
protobuf==4.25.4
pycparser==2.22
pymongo==4.8.0
PySocks==1.7.1
scikit-learn==1.5.1
scipy==1.14.0
selenium==4.22.0
setuptools==65.5.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sympy==1.12.1
tbb==2021.13.0
tensorboard==2.17.0
tensorboard-data-server==0.7.2
threadpoolctl==3.5.0
torch==2.3.1
torchvision==0.18.1
tqdm==4.66.4
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.2
websocket-client==1.8.0
Werkzeug==3.0.3
wsproto==1.2.0
